LayoutTests/html5lib/resources/pending-spec-changes.dat

#data
Placeholder
#errors
#document
| <html>
|   <head>
|   <body>
|     "Placeholder"
